Per CR suggetions: remove from transactions_trace() from SDK, and make into a test utility function
